refactor: rename Auth card
