+++ packages/api-client/src/components/ApiClient/Request/Request.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { useRequestStore } from '../../../stores/requestStore'
-import RequestAuth from './RequestAuth.vue'
+import RequestAuthentication from './RequestAuthentication.vue'
 import RequestBody from './RequestBody.vue'
 import RequestCookies from './RequestCookies.vue'
 import RequestHeaders from './RequestHeaders.vue'
@@ -26,7 +26,7 @@ const { activeRequest, readOnly } = useRequestStore()
     </div>
     <div>
       <RequestVariables :variables="activeRequest.variables" />
-      <RequestAuth />
+      <RequestAuthentication />
       <RequestCookies :cookies="activeRequest.cookies" />
       <RequestHeaders :headers="activeRequest.headers" />
       <RequestQuery :queries="activeRequest.query" />
